Ajout de sécurités pour ne pas supprimer les infos systèmes en base de données
git-svn-id: https://svn.abls-habitat.fr/repo/Watchdog@2586 9f22af27-4b39-4cfe-9f85-4bed9fef4693

+++ trunk/Watchdogd/Message.c
@@ -45,6 +45,8 @@
     gboolean retour;
     struct DB *db;
 
+    if (msg->id < 10000) return(FALSE);
+
     g_snprintf( requete, sizeof(requete),                                                  /* Requete SQL */
                 "DELETE FROM %s WHERE id=%d", NOM_TABLE_MSG, msg->id );
